refactored firebase calls as middleware
